+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `user_data` field is added to the `astarte_device_config_t` and `astarte_device_*_event_t`
   structs. This field can be used to pass user data to callback functions.
+- Utility function `astarte_err_to_name`. Can be used to obtain a string representation of any error
+  returned from the device.
 
 ### Changed
 - Return value of `uuid_generate_v5` and `astarte_hwid_encode` functions from `void` to